Phpunit/repairs

+++ app/Http/Controllers/DocumentsController.php
@@ -11,7 +11,7 @@ use App\Models\User;
 use App\Services\Storage\GetStorageProvider;
 use Illuminate\Http\Request;
 use Ramsey\Uuid\Uuid;
-use Session;
+use Illuminate\Support\Facades\Session;
 
 class DocumentsController extends Controller
 {
@@ -27,8 +27,13 @@ class DocumentsController extends Controller
 
     public function view($external_id)
     {
-        // Eager load the source and nested client relationships to avoid N+1 queries
-        $document = Document::with(['sourceable', 'sourceable.client'])->whereExternalId($external_id)->first();
+        // Eager load the source and conditionally load client for assignable types
+        $document = Document::with(['source' => function ($query) {
+            // Only eager load client for types that have a client relationship
+            if (in_array(get_class($query->getModel()), self::ASSIGNABLE_TYPES)) {
+                $query->with('client');
+            }
+        }])->whereExternalId($external_id)->first();
 
         if (! $document) {
             abort(404);
@@ -36,6 +41,10 @@ class DocumentsController extends Controller
 
         // Check if user has permission to view document via source ownership
         if (! $this->canAccessDocument($document)) {
+            if (request()->expectsJson()) {
+                abort(403, __('You do not have permission to view this document'));
+            }
+
             session()->flash('flash_message_warning', __('You do not have permission to view this document'));
 
             return redirect()->back();
@@ -43,6 +52,7 @@ class DocumentsController extends Controller
 
         $fileSystem = GetStorageProvider::getStorage();
         $file = $fileSystem->view($document);
+
         if (! $file) {
             session()->flash('flash_message_warning', __('File does not exists, make sure it has not been moved from dropbox (:path)', ['path' => $document->path]));
 
@@ -57,8 +67,13 @@ class DocumentsController extends Controller
 
     public function download($external_id)
     {
-        // Eager load the source and nested client relationships to avoid N+1 queries
-        $document = Document::with(['sourceable', 'sourceable.client'])->whereExternalId($external_id)->first();
+        // Eager load the source and conditionally load client for assignable types
+        $document = Document::with(['source' => function ($query) {
+            // Only eager load client for types that have a client relationship
+            if (in_array(get_class($query->getModel()), self::ASSIGNABLE_TYPES)) {
+                $query->with('client');
+            }
+        }])->whereExternalId($external_id)->first();
 
         if (! $document) {
             abort(404);
@@ -66,6 +81,10 @@ class DocumentsController extends Controller
 
         // Check if user has permission to download document via source ownership
         if (! $this->canAccessDocument($document)) {
+            if (request()->expectsJson()) {
+                abort(403, __('You do not have permission to download this document'));
+            }
+
             session()->flash('flash_message_warning', __('You do not have permission to download this document'));
 
             return redirect()->back();
@@ -140,14 +159,17 @@ class DocumentsController extends Controller
      */
     public function uploadToTask(Request $request, $external_id)
     {
-        /**   if (!auth()->user()->can('image-upload')) {
-        session()->flash('flash_message_warning', __('You do not have permission to upload images'));
-        return redirect()->route('tasks.show', $task->external_id);
-        }**/
+        if (! auth()->user()->can('task-upload-files')) {
+            session()->flash('flash_message_warning', __('You do not have permission to upload files'));
+
+            return redirect()->back();
+        }
+
         $task = Task::whereExternalId($external_id)->first();
 
         if (! $task) {
             session()->flash('flash_message_warning', __('Task not found'));
+
             return redirect()->back();
         }
 
@@ -186,7 +208,7 @@ class DocumentsController extends Controller
         }
         Session::flash('flash_message', __('File successfully uploaded'));
 
-        return $task->external_id;
+        return response()->json(['external_id' => $task->external_id], 200);
     }
 
     /**
@@ -195,14 +217,17 @@ class DocumentsController extends Controller
      */
     public function uploadToProject(Request $request, $external_id)
     {
-        /**   if (!auth()->user()->can('image-upload')) {
-        session()->flash('flash_message_warning', __('You do not have permission to upload images'));
-        return redirect()->route('tasks.show', $task->external_id);
-        }**/
+        if (! auth()->user()->can('project-upload-files')) {
+            session()->flash('flash_message_warning', __('You do not have permission to upload files'));
+
+            return redirect()->back();
+        }
+
         $project = Project::whereExternalId($external_id)->first();
 
         if (! $project) {
             session()->flash('flash_message_warning', __('Project not found'));
+
             return redirect()->back();
         }
 
@@ -243,7 +268,7 @@ class DocumentsController extends Controller
         }
         Session::flash('flash_message', __('File successfully uploaded'));
 
-        return $project->external_id;
+        return response()->json(['external_id' => $project->external_id], 200);
     }
 
     public function destroy($external_id)
@@ -253,16 +278,19 @@ class DocumentsController extends Controller
 
             return redirect()->route('tasks.show', $external_id);
         }
-        $fileSystem = GetStorageProvider::getStorage();
 
         $document = Document::whereExternalId($external_id)->first();
-        $deleted = $fileSystem->delete($document);
-        if (! $deleted) {
-            Session()->flash('flash_message_warning', __("Something wen't wrong, we can't find the file on the cloud. But worry not, we delete what we know about the image"));
-        } else {
-            Session()->flash('flash_message', __('File has been deleted'));
+
+        if (! $document) {
+            session()->flash('flash_message_warning', __('Document not found'));
+
+            return redirect()->back();
         }
+
+        // Observer will handle file deletion
         $document->delete();
+
+        session()->flash('flash_message', __('File has been deleted'));
 
         return redirect()->back();
     }
@@ -305,7 +333,7 @@ class DocumentsController extends Controller
         $user = auth()->user();
 
         // Use the morphTo relationship to get the source model
-        $source = $document->sourceable;
+        $source = $document->source;
 
         if (! $source) {
             return false;
